// src/generated/RecordText.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RecordInfo } from "./RecordInfo";

/**
 * A text record with its body.
 */
export type RecordText = { info: RecordInfo, 
/**
 * The frontmatter text between the `---` lines.
 */
frontmatter: string, body: string, };